+++ app/controllers/users/personals_controller.rb
@@ -1,6 +1,4 @@
 class Users::PersonalsController < ApplicationController
-  include ProfileImageUploader
-
   def new
     @user = PersonalUser.new
   end
@@ -11,7 +9,9 @@ class Users::PersonalsController < ApplicationController
 
   def create
     @user = PersonalUser.new(user_params)
+    @image = @user.build_profile_image(image: params[:profile_image].read)
     if @user.save
+      flash[:info] = "プロフィール画像の設定に失敗しました" unless @image.save
       UserMailer.account_activation(@user).deliver_now
       flash[:success] = "仮登録が完了しました。メールから本登録をお願い致します。"
       redirect_to new_users_personal_url
